klientctl: rework ssh package and make it business logic agnostic
